+++ Scripts/CommonServerPython/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
-
+- Added ***BaseClientSession*** - a child of BaseClient class that uses session in order to maintain an open session 
+with the server.
 
 ## [19.10.1] - 2019-10-15
  - Added ***is_debug_mode*** wrapper function for checking if **debug-mode** is enabled.